[機能追加]Ball Calibration UIをweb_debuggerに統合
## 背景
Ball Calibration UIが別パッケージ・別ポート(8095)で提供されており、
デバッグ導線と運用が分断されていたため。

## 変更内容
- crane_ball_calibration_uiを廃止し、実装をcrane_web_debuggerへ移管
- 統合HTTPサーバ(FastAPI)を追加し、/ball-calibration/ と API を同一ポート化
- ポータル導線を外部リンクから内部パスへ変更
- データ再読込時に手動kick overrideをリセットする不具合を修正
- パッケージ/ドキュメント構成を統合後の形に更新
