--- crane_ball_calibration_ui/crane_ball_calibration_ui/__init__.py
@@ -1,1 +1,0 @@
-"""crane_ball_calibration_ui: ボール物理モデルパラメータのインタラクティブ最適化WebUI."""
