--- crane_ball_calibration_ui/package.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>crane_ball_calibration_ui</name>
-  <version>1.0.293</version>
-  <description>
-    ボール物理モデルとキッカーモデルのパラメータをHuman-in-the-Loopで
-    インタラクティブに最適化するためのWebインターフェース
-  </description>
-  <maintainer email="ibis.ssl.team@gmail.com">ibis-ssl</maintainer>
-  <license>MIT</license>
-
-  <exec_depend>python3-fastapi</exec_depend>
-  <exec_depend>python3-numpy</exec_depend>
-  <exec_depend>python3-pydantic</exec_depend>
-  <exec_depend>python3-yaml</exec_depend>
-  <exec_depend>rclpy</exec_depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>crane_lint_common</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
