--- crane_ball_calibration_ui/setup.cfg
@@ -1,7 +1,0 @@
-[develop]
-script_dir=$base/lib/crane_ball_calibration_ui
-[install]
-install_scripts=$base/lib/crane_ball_calibration_ui
-
-[tool:pytest]
-junit_suite_name = crane_ball_calibration_ui
