+++ crane_web_debugger/package.xml
@@ -18,6 +18,11 @@
   <depend>nlohmann-json-dev</depend>
   <depend>rclcpp</depend>
   <depend>std_msgs</depend>
+  <exec_depend>python3-fastapi</exec_depend>
+  <exec_depend>python3-numpy</exec_depend>
+  <exec_depend>python3-pydantic</exec_depend>
+  <exec_depend>python3-uvicorn</exec_depend>
+  <exec_depend>python3-yaml</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>crane_lint_common</test_depend>
