+++ crane_web_debugger/web/ball-calibration/index.html
@@ -6,7 +6,7 @@
   <title>Ball Calibration UI - ibis-ssl</title>
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
-  <link rel="stylesheet" href="/static/style.css">
+  <link rel="stylesheet" href="/ball-calibration/static/style.css">
 </head>
 <body>
 
@@ -318,6 +318,6 @@
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 <script src="https://cdn.plot.ly/plotly-2.27.0.min.js"></script>
-<script src="/static/app.js"></script>
+<script src="/ball-calibration/static/app.js"></script>
 </body>
 </html>
