+++ crane_web_debugger/web_server/ball_calibration/__init__.py
@@ -1,0 +1,1 @@
+"""Ball calibration backend modules for crane_web_debugger."""
